+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - (Feature) (ML) Handlers
 - (Feature) Add P0 Compare Func
 - (Bugfix) Handle optional taints for Storage Operator
+- (Bugfix) Fix Early Connections for 3.10+
 
 ## [1.2.35](https://github.com/arangodb/kube-arangodb/tree/1.2.35) (2023-11-06)
 - (Maintenance) Update go-driver to v1.6.0, update IsNotFound() checks